# OPT: Bypass `submodule add` when adding subdatasets