Instead go manual, avoiding various checks, because we already
know inside `save` that a particular directory is considered untracked
by Git.

This approach is substantially faster. Already with 50 subdatasets
the speed-up is 2x.

Quick'n'dirty benchmark:

```
% rm -rf * .git* .datalad; datalad create . && \
  ( for i in $(seq 50); do datalad create $i; done ) && \
  time datalad save && datalad status
```

Current master

```
datalad save  3.70s user 1.50s system 103% cpu 4.998 total
```

This PR

```
datalad save  1.68s user 1.02s system 108% cpu 2.499 total
```

It may well be that it needs more safety checks that would eat up that
margin. But in the context of
https://github.com/datalad/datalad/issues/2880#issuecomment-667041918
(save thousands of new subdatasets) even much smaller benefits would
still translate to substantial absolute speed-ups.

Thx again to @kyleam for digging out that `update-index` trick.

TODO:

- [x] consolidate config and `.gitmodules` manipulation into a single action for all added submodules at once
- [x] RF helpers into documented standalone functions with tests