@mih 2020-08-03
This works now. The only test failure is:

```
======================================================================
1649ERROR: datalad.distribution.tests.test_install.test_install_recursive
1650----------------------------------------------------------------------
1651Traceback (most recent call last):
1652  File "/home/travis/virtualenv/python3.5.6/lib/python3.5/site-packages/nose/case.py", line 134, in run
1653    self.runTest(result)
1654  File "/home/travis/virtualenv/python3.5.6/lib/python3.5/site-packages/nose/case.py", line 152, in runTest
1655    test(result)
1656AssertionError: Test was too slow (took 31.0307s, threshold was 30.0000s)
```

it is unclear to me how it can get slower (I have not yet observed a scenario where things got slower), but there are some optimization that can still be done.