@mih 2020-08-03
Update: Consolidation on applying all `.gitmodules` and `.git/config` changes at once speeds things up by another 40%, leading to a 3x overall speed-up for 50 subdatasets.